requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sarvamai
qdrant-client
google-genai
httpx
requests
psycopg2-binary
sqlalchemy
